requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Django==4.2
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
hiredis==2.2.2
Pillow==9.5.0
psycopg==3.1.9
psycopg-binary==3.1.9
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
sqlparse==0.4.4
typing_extensions==4.6.0
